Safari-609.1.11.1/LayoutTests/css3/flexbox/flexbox-height-with-overflow-auto.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.flexbox {
    border: 5px solid green;
    position: relative;
    width: 50px;
}

.inline-flexbox {
    border: 5px solid green;
    position: relative;
    height: 50px;
}

.overflow {
    border: 1px solid red;
    overflow: auto;
    min-width: 0;
    min-height: 0;
}

.vertical {
    writing-mode: vertical-rl;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox, .inline-flexbox')">
<div id=log></div>
<div class="flexbox" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox row-reverse" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox vertical" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox row-reverse vertical" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox" data-expected-height="47">
  <div class="overflow align-self-baseline"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox row-reverse" data-expected-height="47">
  <div class="overflow align-self-baseline"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox vertical" data-expected-height="32">
  <div class="overflow align-self-baseline"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="flexbox row-reverse vertical" data-expected-height="32">
  <div class="overflow align-self-baseline"><div style="width: 100px; height: 20px"></div></div>
</div>

<div class="inline-flexbox column" data-expected-width="47">
  <div class="overflow"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column-reverse" data-expected-width="47">
  <div class="overflow"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column vertical" data-expected-width="47">
  <div class="overflow"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column-reverse vertical" data-expected-width="47">
  <div class="overflow"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column" data-expected-width="47">
  <div class="overflow align-self-baseline"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column-reverse" data-expected-width="47">
  <div class="overflow align-self-baseline"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column vertical" data-expected-width="32">
  <div class="overflow align-self-baseline"><div style="width: 20px; height: 100px"></div></div>
</div>

<div class="inline-flexbox column-reverse vertical" data-expected-width="32">
  <div class="overflow align-self-baseline"><div style="width: 20px; height: 100px"></div></div>
</div>
</body>
</html>
