Safari-609.1.11.1/LayoutTests/css3/flexbox/flexbox-height-with-overflow-auto-expected.txt


PASS .flexbox, .inline-flexbox 1 
PASS .flexbox, .inline-flexbox 2 
FAIL .flexbox, .inline-flexbox 3 assert_equals: 
<div class="inline-flexbox column-reverse" data-expected-width="47">
  <div class="overflow align-self-baseline"><div style="width: 20px; height: 100px"></div></div>
</div>
width expected 47 but got 32
FAIL .flexbox, .inline-flexbox 4 assert_equals: 
<div class="inline-flexbox column" data-expected-width="47">
  <div class="overflow align-self-baseline"><div style="width: 20px; height: 100px"></div></div>
</div>
width expected 47 but got 32
PASS .flexbox, .inline-flexbox 5 
PASS .flexbox, .inline-flexbox 6 
FAIL .flexbox, .inline-flexbox 7 assert_equals: 
<div class="inline-flexbox column-reverse" data-expected-width="47">
  <div class="overflow"><div style="width: 20px; height: 100px"></div></div>
</div>
width expected 47 but got 32
FAIL .flexbox, .inline-flexbox 8 assert_equals: 
<div class="inline-flexbox column" data-expected-width="47">
  <div class="overflow"><div style="width: 20px; height: 100px"></div></div>
</div>
width expected 47 but got 32
PASS .flexbox, .inline-flexbox 9 
PASS .flexbox, .inline-flexbox 10 
PASS .flexbox, .inline-flexbox 11 
PASS .flexbox, .inline-flexbox 12 
FAIL .flexbox, .inline-flexbox 13 assert_equals: 
<div class="flexbox row-reverse vertical" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>
height expected 47 but got 32
FAIL .flexbox, .inline-flexbox 14 assert_equals: 
<div class="flexbox vertical" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>
height expected 47 but got 32
PASS .flexbox, .inline-flexbox 15 
PASS .flexbox, .inline-flexbox 16 
 
 
 
 
 
 
 
